+++ src/libraries/dispatcher/PopulationConcentration.h
@@ -50,7 +50,7 @@ public:
 
     virtual unsigned int GetDeadHumans() const;
     virtual unsigned int GetLivingHumans() const;
-    virtual unsigned int GetDensity() const;
+    virtual float        GetDensity() const;
     virtual QString      GetAttitude() const;
 
     virtual const Common::MsgCoordLatLong& GetPosition() const;
